fix: add new func to resolve lint issue
