+++ src/pytemplate/utils/common.py
@@ -2,8 +2,6 @@
 
 from marshmallow import ValidationError
 
-from pytemplate.service.burger import CheeseBurgerBuilder, ChickenBurgerBuilder, VeggieBurgerBuilder
-
 
 def get_choice_input():
     valid_choices = {"chicken", "vegan", "cheese"}
@@ -14,14 +12,3 @@ def get_choice_input():
         return choice
     except (EOFError, KeyboardInterrupt) as err:
         raise ValidationError("Input was interrupted. Please try again.") from err
-
-
-def create_burger(choice):
-    if choice == "chicken":
-        return ChickenBurgerBuilder().bread("Sesame Seed Bun").patty("Chicken").sauce("Ketchup").toppings(["Lettuce", "Tomato"]).build()
-    elif choice == "cheese":
-        return CheeseBurgerBuilder().bread("Sesame Seed Bun").patty("Beef").sauce("Ketchup").toppings(["Lettuce", "Tomato"]).build()
-    elif choice == "vegan":
-        return VeggieBurgerBuilder().bread("Sesame Seed Bun").patty("Veggie").sauce("Ketchup").toppings(["Lettuce", "Tomato"]).build()
-    else:
-        raise ValidationError("Invalid choice!")
